Revert "Issue #11729: Select Previous/Next Layers commands only select one layer."
This reverts commit 31df873d4c9b79fb64172d6fffa4367191f71593.

My commit was not wrong per-se (it didn't add any bug), but it was in
fact not useful either, and only adding a useless list copy for no good
reason.
I thought that action_select_object() was actually selecting objects,
but looking at implementation, it in fact doesn't. It only returns a
pointer to the object to select.