Issue #11950: lua version can be on stderr.
Though the output of `lua -v` is on stdout and was also on stdout by
default, even back with version 5.1, it turns out that lua's print
function was considered configurable back then. And this is very likely
why we now have this report of someone whose lua is on stderr.

So let's still look on stdout, then fallback to stderr it the format
doesn't match.