app: fix pasting into a channel.
It was made impossible to paste into a channel, as this was creating a
normal layer instead. We need to use floating selections when the pasted
target is a GimpChannel type.

Note that the UX is very ugly right now for these (except for the layer
mask case as I improved this a lot with commits 04810ec95ef and
28457ce337), so I opened issue Teams/GIMP/Design/gimp-ux/89 to design
further evolutions.