app: all legacy layer modes are now verified.
Some operations generate slightly different pixels but all differences
are minimal (in worst case, barely 0.3%) and the pixels are perceptually
identical, with basically off-by-one differences (e.g. in some case, I
had pixels whose value when converted to u8 would be 91.4999999…, which
meant that slight float computation different are enough to make it be
either 91 or 92).
Even more, in 2 builds of GIMP 2.10 (flatpak or custom build), I had
either the same result as 2.99 in one of the build, or sometimes a third
result.

But any case, none of these differences warrant being qualified as a
change in the render. Therefore I make the digest to compare to into a
list of digests. Hopefully we won't have to edit this list too often
with more slight variant renders.