icons: simply the build code.
We don't need specific install rules anymore. Looking at commit
73bb302a, I can see why I made one for the second rule of each file: the
gtk-encode-symbolic-svg tool does not support renaming the output file
and in original code, we were using 2 source files depending on the
target size. This changed afterwards, and now the source SVG is always
the right name.

For the first rule, it looks like it was unneeded from the start since
our custom colorsvg2png tool is perfectly able to rename the output
file. Maybe I was just keeping code similar in both rules or that was an
oversight.