gitlab-ci: Do not trigger crossbuilds anymore
febb2126 was proposed temporarely while we had hope of having a fix.
Now, we have !1860 (apparently 99% done) but the main maintainer of
crossbuilds refused to look at it because it is focused on core code
until 3.0 so, with this timeframe, the commit lost its reason to exist.

So, to avoid keeping runners work unnecessarelly with abandoned jobs,
let's not trigger them ('when: never') as I proposed on IRC months ago.